Move base64_{de,en}code to util()
